requirements.txt
# requirements.txt
requests
Pillow
transformers
torch
dotenv